chore: update trace-web metrics sample.
