+++ examples/opentelemetry-web/.eslintrc.js
