+++ examples/opentelemetry-web/examples/fetch/index.html
