+++ examples/opentelemetry-web/examples/fetchXhr/index.html
