+++ examples/opentelemetry-web/examples/fetchXhrB3/index.html
