+++ examples/opentelemetry-web/examples/metrics/index.html
