+++ examples/opentelemetry-web/examples/xml-http-request/index.html
