+++ examples/opentelemetry-web/images/xml-http-request.png
