+++ examples/opentelemetry-web/package.json
@@ -1,14 +1,17 @@
 {
-  "name": "web-tracer-example",
+  "name": "web-opentelemetry-example",
   "private": true,
-  "version": "0.27.0",
-  "description": "Example of using @opentelemetry/sdk-trace-web in browser",
+  "version": "0.30.0",
+  "description": "Example of using @opentelemetry/sdk-trace-web and @opentelemetry/sdk-metrics-base in browser",
   "main": "index.js",
   "scripts": {
     "start": "webpack serve --progress --color --port 8090 --config webpack.dev.config.js --hot --host 0.0.0.0 --compress",
     "start-nc": "webpack serve --progress --color --port 8090 --config webpack.dev.config.js --hot --host 0.0.0.0 --no-compress",
     "start-prod": "webpack serve --progress --color --port 8090 --config webpack.prod.config.js --hot --host 0.0.0.0 --compress",
-    "start-prodnc": "webpack serve --progress --color --port 8090 --config webpack.prod.config.js --hot --host 0.0.0.0 --no-compress"
+    "start-prodnc": "webpack serve --progress --color --port 8090 --config webpack.prod.config.js --hot --host 0.0.0.0 --no-compress",
+    "docker:start": "cd ./docker && docker-compose down && docker-compose up",
+    "docker:startd": "cd ./docker && docker-compose down && docker-compose up -d",
+    "docker:stop": "cd ./docker && docker-compose down"
   },
   "repository": {
     "type": "git",
@@ -17,6 +20,7 @@
   "keywords": [
     "opentelemetry",
     "tracing",
+    "metrics",
     "web"
   ],
   "engines": {
@@ -39,20 +43,20 @@
   },
   "dependencies": {
     "@opentelemetry/api": "^1.0.2",
+    "@opentelemetry/api-metrics": "0.30.0",
     "@opentelemetry/context-zone": "^1.0.1",
     "@opentelemetry/core": "^1.0.1",
+    "@opentelemetry/exporter-metrics-otlp-http": "0.30.0",
+    "@opentelemetry/exporter-trace-otlp-http": "0.30.0",
+    "@opentelemetry/exporter-zipkin": "^1.0.1",
+    "@opentelemetry/instrumentation": "0.30.0",
+    "@opentelemetry/instrumentation-fetch": "0.30.0",
+    "@opentelemetry/instrumentation-xml-http-request": "0.30.0",
+    "@opentelemetry/propagator-b3": "^1.0.1",
+    "@opentelemetry/sdk-metrics-base": "0.30.0",
     "@opentelemetry/sdk-trace-base": "^1.0.1",
     "@opentelemetry/sdk-trace-web": "^1.0.1",
-    "@opentelemetry/exporter-zipkin": "^1.0.1",
-    "@opentelemetry/semantic-conventions": "^1.0.1",
-    "@opentelemetry/propagator-b3": "^1.0.1",
-    "@opentelemetry/resources": "^1.0.1",
-    "@opentelemetry/instrumentation": "0.27.0",
-    "@opentelemetry/instrumentation-fetch": "0.27.0",
-    "@opentelemetry/instrumentation-xml-http-request": "0.27.0",
-    "@opentelemetry/exporter-trace-otlp-http": "0.27.0",
-    "@opentelemetry/exporter-metrics-otlp-http": "0.27.0",
-    "@opentelemetry/sdk-metrics-base": "0.27.0"
+    "@opentelemetry/semantic-conventions": "^1.0.1"
   },
   "homepage": "https://github.com/open-telemetry/opentelemetry-js/tree/main/examples/tracer-web"
 }
