+++ examples/opentelemetry-web/webpack.dev.config.js
