+++ examples/opentelemetry-web/webpack.prod.config.js
