+++ lerna.json
@@ -7,6 +7,7 @@
     "experimental/backwards-compatability/*",
     "integration-tests/*",
     "selenium-tests",
-    "examples/otlp-exporter-node"
+    "examples/otlp-exporter-node",
+    "examples/opentelemetry-web"
   ]
 }
